taipy_app/requirements.txt
pyspark
taipy
synapseml